2023-11-01 수정사항-1 반영

+++ real-estate-domain/src/main/java/io/gig/realestate/domain/realestate/price/PriceInfo.java
@@ -29,7 +29,7 @@ public class PriceInfo extends BaseTimeEntity {
 
     private int depositPrice;
 
-    private int revenueRate;
+    private double revenueRate;
 
     private int averageUnitPrice;
