+++ real-estate-domain/src/main/java/io/gig/realestate/domain/realestate/price/dto/PriceCreateForm.java
@@ -19,7 +19,7 @@ public class PriceCreateForm {
 
     private int depositPrice;
 
-    private int revenueRate;
+    private double revenueRate;
 
     private int averageUnitPrice;
